fix tests #4244

+++ src/Repository/SignalementRepository.php
@@ -549,7 +549,7 @@ class SignalementRepository extends ServiceEntityRepository
                 ->setParameter('partners', $partners);
         }
 
-        return (int) $qb->getQuery()->getSingleScalarResult();
+        return (float) $qb->getQuery()->getSingleScalarResult();
     }
 
     /**
